Fix error in `Version#<=>`.
When comparing versions, our internal `#cmp(Int32?, Int32?)` function
looked like it could return a `nil` to the compiler, which meant that
`#<=>` could return a `nil`, which meant that our comparison operators
from `Comparable` could not be generated. Fix it by altering the logic
of `#cmp()` to avoid the possiblity of a `nil`, and change the function
signature to reflect this fact. This makes the compiler happy, allowing
`Comparable` to be happy.

We reinforce this by adding some `Comparable` checks to our comparison
spec in the tests.