Get list of outdated versions.
Outdated versions are found for each plugin provided to
`Outdated::run()`. We find the outdated version (and its upgrade) by
seeing if there are any newer versions available, and then taking the
closest one.

We would like to eliminate versions for which we already have a version
installed from the same series, but that's a harder problem.