Tweak output to include non-outdated versions.
Output all versions, with an appended `"< [new-version]"` for outdated
versions. This will let us check which of the outdated versions are
really outdated, and which ones already have a new version in the series
installed.

We'd do this automatically, but it's actually a fairly non-trivial
problem figuring out whether we want to ignore minor versions or not.
This gets us close enough that it saves us some time, which is the whole
point of the exercise.